Pull out ABY
